rename storage to preferences
